feat(ray-so): add custom title support for RaySo command
This commit introduces the ability to specify a custom title when
generating a code image using the `:RaySo` command. If no title is
provided, the filename is used as the default title. This change
enhances the flexibility of the command and allows users to better
customize their generated code images.

BREAKING CHANGE: The `:RaySo` command now accepts arguments. The
first argument is used as the custom title. If no arguments are
provided, the filename is used as the default title.