feat: Add RaySoClipboard command for copying image to clipboard
This commit introduces a new command `:RaySoClipboard` which generates
a code image and directly copies it to the clipboard. This is achieved
by launching a headless browser, navigating to the generated image URL,
and triggering the copy action. This feature enhances the usability of
the plugin by providing an additional way to use the generated code
images.